requirements.txt
dj-database-url>=3.1.0
django>=6.0.1
django-ckeditor>=6.7.0
django-ratelimit>=4.1.0
python-dotenv>=1.2.1
ratelimit>=2.2.1
supabase>=2.27.1
